move sanitize_url and get_url_group
