+++ corehq/util/metrics/tests/test_utils.py
@@ -4,6 +4,7 @@ from django.test import SimpleTestCase
 from testil import eq
 
 from corehq.util.metrics import make_buckets_from_timedeltas, DAY_SCALE_TIME_BUCKETS, bucket_value
+from corehq.util.metrics.utils import sanitize_url, get_url_group
 from corehq.util.test_utils import generate_cases
 
 
@@ -36,3 +37,25 @@ class MetricsUtilsTest(SimpleTestCase):
 ], MetricsUtilsTest)
 def test_bucket_value(self, value, buckets, unit, expected):
     self.assertEqual(bucket_value(value, buckets, unit), expected)
+
+
+@generate_cases([
+    (
+        '/a/uth-rhd/api/a26f2e21-5f24-48b6-b283-200a21f79bb6/20150922T034026.MP4',
+        '/a/*/api/*/20150922T034026.MP4'
+    ),
+    ('/a/ben/modules-1/forms-2/uuid:abc123/', '/a/*/modules-*/forms-*/uuid:*/')
+], MetricsUtilsTest)
+def test_sanitize_url(self, url, sanitized):
+    self.assertEqual(sanitize_url(url), sanitized)
+
+
+@generate_cases([
+    ('/', 'other'),
+    ('/a/*/api', 'api'),
+    ('/a/domain', 'other'),
+    ('/1/2/3/4', 'other'),
+    ('/a/*/cloudcare', 'cloudcare'),
+], MetricsUtilsTest)
+def test_url_group(self, url, group):
+    self.assertEqual(get_url_group(url), group)
