+++ corehq/util/metrics/utils.py
@@ -1,6 +1,34 @@
+import re
 from datetime import timedelta
 
 from django.conf import settings
+
+WILDCARD = '*'
+
+
+def sanitize_url(url):
+    # Normalize all domain names
+    url = re.sub(r'/a/[0-9a-z-]+', '/a/{}'.format(WILDCARD), url)
+
+    # Normalize all urls with indexes or ids
+    url = re.sub(r'/modules-[0-9]+', '/modules-{}'.format(WILDCARD), url)
+    url = re.sub(r'/forms-[0-9]+', '/forms-{}'.format(WILDCARD), url)
+    url = re.sub(r'/form_data/[a-z0-9-]+', '/form_data/{}'.format(WILDCARD), url)
+    url = re.sub(r'/uuid:[a-z0-9-]+', '/uuid:{}'.format(WILDCARD), url)
+    url = re.sub(r'[-0-9a-f]{10,}', '{}'.format(WILDCARD), url)
+
+    # Remove URL params
+    url = re.sub(r'\?[^ ]*', '', url)
+    return url
+
+
+def get_url_group(url):
+    default = 'other'
+    if url.startswith('/a/' + WILDCARD):
+        parts = url.split('/')
+        return parts[3] if len(parts) >= 4 else default
+
+    return default
 
 
 def make_buckets_from_timedeltas(*timedeltas):
